Refactor: 출력 리스트(하드코딩) 파일 추가 및 호출 방식 변경
